Move js number to elements lib
