+++ lib/tests/streamlit/elements/number_input_test.py
@@ -20,11 +20,11 @@ import pytest
 from parameterized import parameterized
 
 import streamlit as st
+from lib.streamlit.elements.lib.js_number import JSNumber
 from streamlit.errors import (
     StreamlitAPIException,
     StreamlitValueAboveMaxError,
 )
-from streamlit.js_number import JSNumber
 from streamlit.proto.Alert_pb2 import Alert as AlertProto
 from streamlit.proto.LabelVisibilityMessage_pb2 import LabelVisibilityMessage
 from streamlit.proto.NumberInput_pb2 import NumberInput
